Update for TLS 1.3.
This is a very large change. A lot of the text was reorganized to
accomodate both client- and server-offered values. GREASE all the
various extension points added in TLS 1.3, as discussed on the list.

Since TLS 1.3 also has an applicable extension point that we decided not
to add a registry for (versions) or haven't added and haven't decided
whether to (PSK KEs), I've restored the "temporary" number list (still
marked with TBD) in the body of the document, otherwise it never says
what the temporary values are.

For PSK KE values, I borrowed the scheme from
draft-bishop-httpbis-grease-00.